move do_anything_now to extend the base
