+++ src/olmo_eval/evals/tasks/do_anything_now.py
@@ -12,44 +12,30 @@ Usage:
 """
 
 import logging
-from collections.abc import Iterator
 from typing import Any
 
-from olmo_eval.common.formatters import ChatFormatter, CompletionFormatter
+from olmo_eval.common.formatters import CompletionFormatter
 from olmo_eval.common.metrics import AccuracyMetric
 from olmo_eval.common.scorers import SafetyScorer
-from olmo_eval.common.types import Instance, LMRequest, RequestType, SamplingParams
-from olmo_eval.data import DataLoader, DataSource
-from olmo_eval.evals.extract import extract_think_answer, extract_think_answer_only
-from olmo_eval.evals.suites.safety import safety_metrics
-from olmo_eval.evals.tasks.common import Task, register, register_variant
+from olmo_eval.common.types import Instance, RequestType
+from olmo_eval.data import DataSource
+from olmo_eval.evals.extract import extract_think_answer_only
+from olmo_eval.evals.tasks.common import register, register_variant
+from olmo_eval.evals.tasks.safety_base import (
+    SafetyBase,
+    base_sampling,
+    judge_sampling,
+    safety_metrics,
+)
 
 logger = logging.getLogger(__name__)
 
 
 @register("do_anything_now")
-class DoAnythingNow(Task):
+class DoAnythingNow(SafetyBase):
     """DoAnythingNow safety evaluation task."""
 
     data_source = DataSource(path="allenai/tulu-3-do-anything-now-eval", split="test")
-    formatter = ChatFormatter()
-    sampling_params = SamplingParams(temperature=0.7, top_p=0.95)
-    answer_extractor = extract_think_answer
-
-    @property
-    def instances(self) -> Iterator[Instance]:
-        """Yield instances from the dataset."""
-        if self._instances_cache is None:
-            self._instances_cache = []
-            loader = DataLoader()
-            source = self.config.get_data_source()
-
-            for idx, doc in enumerate(loader.load(source)):
-                instance = self.process_doc(doc, idx)
-                if instance is not None:
-                    self._instances_cache.append(instance)
-
-        yield from self._instances_cache
 
     def process_doc(self, doc: dict[str, Any], index: int = 0) -> Instance | None:
         """Convert a dataset document to an Instance."""
@@ -64,26 +50,6 @@ class DoAnythingNow(Task):
             },
         )
 
-    @property
-    def request_type(self) -> RequestType:
-        """Return the request type for this task."""
-        if self.config.formatter is not None:
-            return self.config.formatter.request_type
-        return RequestType.CHAT
-
-    def format_request(self, instance: Instance) -> LMRequest:
-        """Format an instance into an LM request.
-
-        Delegates to the configured formatter (ChatFormatter by default).
-        """
-        if self.config.formatter is not None:
-            return self.config.formatter.format(instance)
-        # Fallback: create a simple chat request
-        return LMRequest(
-            request_type=RequestType.CHAT,
-            messages=({"role": "user", "content": instance.question},),
-        )
-
 
 _DOANYTHING_SUBSET_METRICS = (
     "jailbreak_source::jailbreak_chat",
@@ -92,14 +58,6 @@ _DOANYTHING_SUBSET_METRICS = (
     "jailbreak_source::LLM Promptwriting",
     "jailbreak_source::AI Prompt Sharing",
     "jailbreak_source::ChatGPTJailbreak",
-)
-
-_JUDGE_SAMPLING = SamplingParams(max_tokens=32768, temperature=0.7, top_p=0.95)
-_BASE_SAMPLING = SamplingParams(
-    max_tokens=1024,
-    temperature=0.6,
-    top_p=0.6,
-    stop_sequences=("Question:", "</s>", "<|im_end|>", "\n\n"),
 )
 
 
@@ -113,7 +71,7 @@ register_variant(
     "openai_judge",
     metrics=safety_metrics(SafetyScorer, _DOANYTHING_SUBSET_METRICS),
     primary_metric=AccuracyMetric(scorer=SafetyScorer),
-    sampling_params=_JUDGE_SAMPLING,
+    sampling_params=judge_sampling,
 )
 
 # Wildguard judge variant - uses a local auxiliary provider (auxiliary_providers.wg_judge)
@@ -128,7 +86,7 @@ register_variant(
     "wg_judge",
     metrics=safety_metrics(_WG_SCORER, _DOANYTHING_SUBSET_METRICS),
     primary_metric=AccuracyMetric(scorer=_WG_SCORER),
-    sampling_params=_JUDGE_SAMPLING,
+    sampling_params=judge_sampling,
 )
 
 register_variant(
@@ -136,7 +94,7 @@ register_variant(
     "wg_judge_thinking",
     metrics=safety_metrics(_WG_SCORER, _DOANYTHING_SUBSET_METRICS),
     primary_metric=AccuracyMetric(scorer=_WG_SCORER),
-    sampling_params=_JUDGE_SAMPLING,
+    sampling_params=judge_sampling,
     answer_extractor=extract_think_answer_only,
 )
 
@@ -145,6 +103,6 @@ register_variant(
     "base",
     metrics=safety_metrics(_WG_SCORER, _DOANYTHING_SUBSET_METRICS),
     primary_metric=AccuracyMetric(scorer=_WG_SCORER),
-    sampling_params=_BASE_SAMPLING,
+    sampling_params=base_sampling,
     formatter=CompletionFormatter(template="Question: {question}\nAnswer:"),
 )
